requirements.txt
# Core data science
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
scipy==1.13.1

# HTTP / API
requests==2.32.3
requests-cache==1.2.1

# Database
sqlalchemy==2.0.31
alembic==1.13.2

# Configuration
python-dotenv==1.0.1
pyyaml==6.0.1

# MLOps
mlflow==2.15.1

# Drift detection / monitoring
evidently==0.4.33

# Visualization
matplotlib==3.9.1
plotly==5.23.0
seaborn==0.13.2

# Dashboard
streamlit==1.37.1

# Calendars / holidays
holidays==0.94

# Notebook
jupyter==1.0.0
jupyterlab==4.2.4
ipykernel==6.29.5

# Data serialization
pyarrow==17.0.0
fastparquet==2024.5.0

# Gradient boosting / tree models
catboost>=1.2

# Testing
pytest==8.3.2
pytest-cov==5.0.0
pytest-mock==3.14.0

# Code quality
ruff==0.5.6
mypy==1.11.1

# Job scheduling (for drift monitor cron)
apscheduler==3.10.4

# Progress bars
tqdm==4.66.5

# Utilities
joblib==1.4.2
click==8.1.7
tabulate==0.9.0
